+++ mbsimgui/mbsimgui/force.cc
@@ -1,6 +1,6 @@
 /*
-    MBSimGUI - A fronted for MBSim.
-    Copyright (C) 2026 Martin Förg
+   MBSimGUI - A fronted for MBSim.
+   Copyright (C) 2026 MBSim-Env
 
   This library is free software; you can redistribute it and/or 
   modify it under the terms of the GNU Lesser General Public 
@@ -15,22 +15,16 @@
   You should have received a copy of the GNU Lesser General Public 
   License along with this library; if not, write to the Free Software 
   Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301 USA
-*/
+   */
 
-#ifndef _TRANSMISSION__H_
-#define _TRANSMISSION__H_
+#include <config.h>
+#include "force.h"
+#include "objectfactory.h"
 
-#include "link_.h"
-#include "link_property_dialog.h"
+using namespace std;
 
 namespace MBSimGUI {
 
-  class GeneralizedTransmission : public DualRigidBodyLink {
-    MBSIMGUI_OBJECTFACTORY_CLASS(GeneralizedTransmission, DualRigidBodyLink, MBSIM%"GeneralizedTransmission", "Generalized transmission");
-    public:
-      PropertyDialog* createPropertyDialog() override { return new GeneralizedTransmissionPropertyDialog(this); }
-  };
+  MBSIMGUI_REGOBJECTFACTORY(GeneralizedForce);
 
 }
-
-#endif